iti php folder
This PHP folder provides a comprehensive solution for building a customer registration form with CRUD (Create, Read, Update, Delete) operations, along with the ability to display all customer data in a table format. The code is designed to be easily integrated into your project, facilitating efficient management of customer information.